+++ src/lib_cli_exit_tools/__init__conf__.py
@@ -1,12 +1,12 @@
 # CONF
 
-name = 'lib_cli_exit_tools'
-title = 'functions to exit an cli application properly'
-version = 'v0.0.1'
-url = 'https://github.com/bitranox/lib_cli_exit_tools'
-author = 'Robert Nowotny'
-author_email = 'bitranox@gmail.com'
-shell_command = 'lib_cli_exit_tools'
+name = "lib_cli_exit_tools"
+title = "Functions to exit a CLI application properly"
+version = "0.1.0"
+url = "https://github.com/bitranox/lib_cli_exit_tools"
+author = "Robert Nowotny"
+author_email = "bitranox@gmail.com"
+shell_command = "lib_cli_exit_tools"
 
 
 def print_info() -> None:
@@ -14,7 +14,7 @@ def print_info() -> None:
 
 Info for {name}:
 
-    functions to exit an cli application properly
+    Functions to exit a CLI application properly
 
     Version : {version}
     Url     : {url}